- moved getRelationData to util function
